requirements.txt
Django==3.2.4
django-cors-headers==4.2.0
environs==9.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-simple-history==3.0.0
psycopg2-binary==2.9.6
python-decouple==3.3
